WIP: Get some stubbs set up for streaming
Just jotting down the first steps in streaming a serialized cJSON
object. The point being that instead of allocating and/or reallocating
a potentially large buffer, the serialized data just gets periodically
written out. It's similar to the printbuffer, but instead of
reallocating when the buffer is full, it just flushes the current string
out to a user provided callback.