Flesh out all of the stream functions
Still need to handle strings larger than the stream buffer and some kind
of error handling.

Still need to fix tab depth for objects and arrays.

Also need to consider adding a function for simply adding characters to
the stream since that would be a simpler to use for the array and object
functions.