Include path compatibility with shared library
Some project may wish to write code that can be compiled against cJSON
come from either the system shared library, or by including the cJSON
git repository as a submodule (in essence, the same as copying the
source into the project manually). However, this leads to a
compatibility problem: when using the system library, cJSON.h must be
included as:

 #include <cjson/cJSON.h>

whereas if it is included locally, it must be included as:

 #include <cJSON.h>

This makes it difficult to write code that doesn't care which method is
used to provide cJSON.

To remedy this, the cJSON headers are moved to an "include" directory
with the proper path, and the source is fixed up to use new include
path.