adding string length to string objects
there is an isssue with the strings in this library.
on one hand when parsing the string it's parsed from \" till \"
on the other hand when you want to actually get the data it's aviable
but if you had something like this: "data\u0000\n" it's a problem since
the \0 is before the end of the string.
To solve the zero teminrated string i added an extra value
stringvalue_len that can be retrived using a function that holds the
length of the pased string. The length takes into account the parsing of
special escaped chars.