Use 'strcmp' to compare string literals
Using 'strncpy' offers the potential for:
 * Coding errors when the string is changed but the size is forgotten
 * Incorrect result if the variable starts with the string literal, but
   also contains additional characters

On the other hand, strcpy with a string literal is guaranteed to be
bounded, and will always compare up to the first null character.

The possibility that the variable string memory is smaller than the
string literal still exists, but if the code has unterminated strings
running around there are other problems.