Adjust the size of cJSON from 64 to 56 through memory alignment
By adjusting the position, the valueint and type
memory are aligned, occupying a total of 8 bytes
instead of 16 bytes, so sizeof(cJSON) is reduced
from 64 bytes to 56 bytes.

p.s. Although this optimization may not be
effective in the memory allocator's view, because
the size that the allocator may allocate is 64 bytes,
it may be effective in some scenarios, and it will
not be worse than the current situation.